Change the way counters are turned into slices.
Given a counter with value v0 and v1 at time t0 and t1, respectively:
the current implementation turns this into a slice [t0, t1) with v0.

With this change it'll be (t0, t1] v1, which more closely matches
what the counters represent (esp. GPU counters). A counter value at
some time t, represents what the counter *was* up to time t, not what
it it will be right after time t.

For example, a "triangles processed" counter shows the triangles that
were processed in the previous time period, not the following, when
sampled.