Add api.Cmd.Terminated(), propagate it to UI (#3404)
This new method identifies whether a command has properly terminated or not.

Mutating a command that has not terminated results in applying pre-fence observation, and then aborting (api.Abort()) at the fence. An extra argument is added to api.ForeachCmd to filter out non-terminated commands, or not. The only place where we actually want to consider the non-terminated commands is when the UI CommandTree is generated.

In the CommandTree view, non-terminated commands are marked in the UI by stating that they did not terminate.

This was tested on a few apps, including one with heavy concurrency such that at the end of a capture some commands in other threads are non-terminated, and another where we forced a crash in a VkQueueSubmit by passing invalid data for arguments related to semaphores.