Fix "Capturing" window title: could be any type of trace
The window title use to be "Capture Graphics Trace", which is
incorrect when a system trace is being captured. This fix changes it
to the generic "Capturing...".

It would be possible to have different window title based on which
type of trace is being captured, but I feel this adds unecessary
complexity for an additional information which is probably not worth
it. If the user does not remember which type of capture has been
started, the capture output file name already gives a hint.