Improve device selection (#3537)
The "serial" of the selected trace device is saved and used to automatically
fill the device drop-down when taking a new trace. This only works with Android
devices but not with Stadia as a serial is not available. Fall back to using the
device's/instance's (friendly) name when serial is not available.
Also, extend to Stadia the logic for which if two devices are available and
exactly one is an Android device/Stadia instance, then that one should
automatically be selected if no previous selection is available.

Bug: b/145499378