Non mec looping (#3549)
* Patch up Frame Looping transform so it works with EOS

The FrameLooping transform invalidated the api state for all transforms
following it. This is a shortcoming of the transform system which was
not designed to be aware of control flow. This change patches up the
Frame Looping transform so we can avoid the problem for now. I am going
to follow this commit with a design doc for a revised transform system
that will not have this problem (and require this patch).

* Make FrameLooping robust to NoCmdID inside loop