Refactor and expose PNG-writing functionality
This change refactors the PNG-writing functionality in the virtual
swapchain layer so that the PNG data is passed to a function (provided
as a function pointer), rather than directly written to a file.  When
invoked directly from the layer, a function that does indeed write the
image to a file is used.  This more general PNG-writing facility is
exposed to client code as an extern C function.