Add vulkan command-buffer-splitter.
This splits the vulkan command-buffer at a given set of subcommands.
It drops the command-buffer out of a renderpass, since
the operations allowed in a renderpass are limited. It then inserts
a CommandBufferInsertionCommand into the current command-buffer.

It then re-enters a renderpass (as necessary) and continues on its way.
Furthermore it inlines all secondary command-buffers that are found
in the command-buffer that must be split. This also allows the same
functionality.

This is the first commit on the way to being able to batch replays
for any requests that need data in the middle of a command-buffer.
Future PRs will come to expand on this functionality.

BUG: <none>