Crash early upon cmd.Mutate() error in all MutateAndWrite()
If cmd.Mutate() leads to an error, then we cannot trust the state to
be valid anymore. MutateAndWrite() wraps calls to cmd.Mutate(), this
change makes sure to crash early (here, using log.F()) upon any
cmd.Mutate() error.

We can think of propagating the error above MutateAndWrite(). Yet,
MutateAndWrite() is part of the TransformWriter interface, and there
is no obvious way to enforce developers working on Transforms to
properly handle possible errors returned by MutateAndWrite() in the
logic of their transform.

As a middle ground, we choose to have MutateAndWrite() crash early
upon any cmd.Mutate(). Again, we cannot strictly enforce that, it is
just documented in the interface comment.

Bug: b/145003736