Make sure Perfetto daemons are running.
Perfetto daemons are not required to run on the device, and the daemons must be
running in order to allow the data producers register themselves, hence before
we do any interaction between the tool and Perfetto, we must make sure Perfetto
daemons are running and persistent.

Note that there's a delay between setting the system property and the daemons
finish starting, hence we need to set the system property as early as possible.

Also update the test faked device to Android 10.

Bug: b/148566762