Simplified queued command tracking.
This removes an edge-case that is not quite valid in the spec
but that we handled anyway. This significantly simplifies
the queued command tracking code.

It removes the repsonsibility of the API itself to handle
debug markers (that was a server detail leaking into the
API).

It also moves all of the tree-building logic into one
place.