Switched out from linear scan in map.inc.
We were using a linear scan for the hash-map. The new
changes made for read_framebuffer started hitting a very bad
edge-case with this. (layer | (level << 16)).

Our hash (for integers) is hash(x)->x. This is 1) very fast,
2) provides good coherency, but when half of your key gets cut
out, we clash a lot, that on top of the linear scan means we
bunch up in the bottom of the map.

This updates our map to use an LCG to determine secondary buckets.
c=1, m = 69069 mod = 2^44. This gives us some good benefits.
1) These paramters allow a won't cycle. (We will go through every bucket
    before starting again)
2) This won't overflow the 64-bit integer we are using.
3) We don't get nearly the same level of clustering.
4) We only do this if the first (cheap) bucket was missed.