Fixes hovers causing loading tracks not to repaint.
There was a race between the renders and the hovers, where if the hover
computation happend before the render of the track, the data request of
the render was deduped against the one from the hover computation,
causing the render repaint callback to be dropped. This fixes this by
combining the callbacks when deduping and invoking all encountered
callbacks, when the data is finally ready.