Merge from upstream:
Trim down the atrace categories.
Updates to the ftrace categories we use.

 - Always listen to the process creation and ending categories to
   accurately track process metadata throughout the trace.
 - Remove unneeded sched/sched_process_exit
 - Add kmem/rss_stat if memory is selected. This used to be implied by
   "memory" atrace category.
Fix how we get process data.

 - Always scan the process data at startup and write it into a
   separate buffer.
 - Always peridically scan the process data and include it in the main
   buffer.

Set the ftrace drain period in the trace config.

Instructs the ftrace producer to drain the kernel ftrace buffer every
250ms. The buffer is 8Mb per CPU, which should be enough for than 1s.

Disable the per-rail power counters.

We don't currently support them in the UI.

Enable the VulkanAPI producer if slices are selected.

Not when counters are selected.

Write the trace into file if it's a long trace.