Delete a swapchain image only if it is not shared
When a swapchain is passed as the oldSwapchain argument of
vkCreateSwapchainKHR(), its images may be reused by the newly created
swapchain, and this can lead to images being shared between
swapchains. Hence, when a swapchain is destroyed, we must delete only
the images that are not being used by another swapchain.

Bug: b/148774659
Test: Manual: capturing Khronos sample AFBC does not crash anymore