Disable alpha channel by default in ImagePanel
In real Vulkan workloads, particularly with swapchain images, there is
often an alpha channel present but ignored, and the app is not careful
about what it writes into the alpha channel as a result. AGI already had
some logic to try and detect when the alpha channel was "unused" based
on contents, but junk rendered into it defeats that heuristic.

Disabling the alpha is a better default.

Bug: b/150420088