Replay profiler support for Mali
This builds on top of Adam's changes to support renderpass-level
linking. Mali is more troublesome in that we don't have a single
toplevel slice for each renderpass -- instead, we may have many slices
for the same renderpass instance scattered across the timeline.

To be completely robust, we need some additional identifier which
increments as we progress through a command buffer, and to have that
identifier included in all the slices originating from that command
buffer. This is the same thing we need in order to accurately attribute
compute slices to CmdDispatch in Adreno.