Change the key type of the ResourceMap type.
Currently the key is the Resource interface (i.e. a pointer). This key
is only useful if the exact same state object is used for map lookups.
Since the map is put into the DB, this is not usually the case.

With this change, the key is the resource handle. However, since the APIs
are allowed to re-use handles, they are no longer "time-globally" unique,
and so the mapping is only valid at a specific point in the trace. It is
currently only used at specific points in the trace, so this is OK.