Reserve mapped memory at state initialization
Previously, if device memory was mapped prior to MEC start, the
allocator was not informed that the memory range was in use. This
led to other memory allocations and/or mappings being placed over
the top.

This would lead to quiet memory corruption in the common case,
or if we got unlucky and placed another memory mapping over the same
range, bizarre explosions when eventually trying to unmap & delete
pre-MEC-start device memory object.

Fixes Swiftshader 'UnmapMemory' crash on Swarming.

Bug: b/151297033