Add DeviceDialog to select replay device (#192)
When zero or more than one replay devices are available for a Graphics trace,
ask the user to select a replay device via a dialog. This replaces
ReplayDeviceSelector.

models.devices.loadReplayDevice() now returns an empty list (and not
null) when the replay device loading succeeded but did return an empty list.
This enables to monitor when the replay devices have been loaded (before,
replayDevices was null if either devices were not loaded yet, or there
was no replay devices available).