Store all images tightly-packed in MEC.
For linear tiled images, we need to 1) share the pool with the bound
memory for this image (to handle aliasing properly) and 2) store the data
in the pool in the linear layout (i.e. with extra padding). However, the
image primer in state reconstruction currently doesn't handle non-
tightly packed data. With this change we may now break certain aliasing
corner-cases, in favor of not breaking linear tiled images.

Once the image primer can correctly extract the data in tiled layout,
this commit can be reverted (plus adding code to ensure the pools are
shared).

Tracked in http://b/154936418