Restart GAPIR only upon a new replay request (#367)
When a ping to the replayer fails, just remove this replayer, do not
proactively restart it. Removing it from the Client replayers list
ensures that it will be restarted upon the Connect() of the next
replay request. This change enables to e.g. start taking a new
trace on a device that was used for replay, without having the
replayer restarting and taking over the app you want to trace.

Bug: b/159997874
Test: manual