Clean up GPU profiling integration with BSP drivers. (#376)
Instead of trying to provide an option to profile against system driver, AGI
should always query and use the latest available drivers on the device. When
developer driver is present, AGI will always try to set up and use it, and when
fail to do so, fall back to use GPU profiling libraries in system image. If
developer driver is not installed or the device doesn't support developer
driver, always attempt to use GPU profiling libraries in system image.

Bug: b/148970509
Test: Perform below commands on supported devices.
Test: bazel run gapit -- device --os android
Test: bazel run gapit -- validate_gpu_profiling --os android