Fix Vulkan handling of mesh requests on groups/non-draw calls.
The Vulkan implementation of the MeshProvider interface was broken and
not returning the correct values when the mesh was requested on a
non-draw call command. This broke the logic in the resolve and would
cause meshes to be returned when they shouldn't and not returned when
they should. This became exasperated by the recent addition of the draw
call grouping, which made it possible to select a range of commands
containing a draw call, which was broken, and thus more obvious.