Basic track filtering feature.
Filters tracks based on track title. If a group's title matches the
search term, all children are included, even if a child may not match.

If a group contains a child that matches, the group itself is shown as
well, but only containing children that match. This is so that context
is not lost.