Force AGI to launch data producer when setup Perfetto tracing. (#423)
During device discovery time AGI pushes a binary to the device to launch the
perfetto data producer for GPU counters, which only gets called once. And hence
any subsequent failure that causes the binary to stop running would lead to
tracing failure. This patch makes sure AGI force the binary to start before
starting a new tracing session. Note that if the driver is changed since device
discovery time, it could lead to undefined behaviour because the driver
behaviour can change and/or the gpu counters are different.

Bug: b/148420473