Emit warnings when replay instance creation fails (#432)
On Android, the replay happens in gapid APK, where any layer/extension
provided by a library in the original APK will not be present. So if
the original app depends on such layers, the instance creation will
fail. Still, we cannot simply return false upon creation failure here,
as the app may also have tried to create an instance and failed, and
will subsequently re-try to create another instance with different
parameters. Yet, instance creation is likely to be expected to be
successful most of the time, we do emit a debug message upon failure.

Bug: b/158671749