[Frame events] Fix view names when layer name starts with a number (#438)
There are cases where some apps do not set layer names properly, and we
receive, for example, "#0" as layer names. After removing
non-alphanumeric characters, the name we will have is just "0". SQL
tables cannot start with a number. A simple fix here is to prepend some
text to all layers. These names are used only internally for view
creation, so the UI is not going to be affected.

Bug: 166511262