[Frame events] Fix instant events depth and negative duration slices (#437)
Due to missing events (happens at the end of the trace and sometimes randomly), the artificial slices we create for the phases track end up having negative duration. Our assumption that <=0 duration is meant for instant events becomes a problem here because of the -1 duration for these slices. This change fixes this by ignoring -1 duration slices altogether, since they do not represent any meaningful data and it goes against our design to draw diamonds and slices together on the same track.

Bug: 159623083