Extend annotation support in vulkan .api files
Add the @str, @structextend, @values, @len, @len_bytes, @returedchain
and @returnedonly annotations to Vulkan API Class, Fields and
Command Parameter types inside the .api files. These annotations
are generated by extracting the information from Khronos vk.xml
registry file.

More information can be found in
https://github.com/KhronosGroup/Vulkan-Docs/blob/master/xml/registry.rnc

@structextends("struct1", "struct2", ...) before classes to specify
the parents of the structure in the pNext chain.

@values(VK_STRUCTURE_TYPE1, VK_STRUCTURE_TYPE2, ...) before
VkStructureType fields with a valid value for the field.

@str before null terminated strings.

@len("member1", "member2", ....) before dynamic arrays with names
of members that contains the length of the array for each dimension.
This will also be applied to command parameters.

@len_bytes("member1", "member2", ...) before dynamic arrays with names
of members that contains the length of the array for each dimension.
The data of this array should be interpreted as an array of bytes.
This will also be applied to command parameters.

@returnedonly before classes that are only filled by API.

@returnedchain before chainable classes that contain pNext and sType
fields. The chained structure is created by the application and the API
fills the rest of the fields.