Avoid recording invalid command buffer (#474)
Some of the command buffers contained in the initial state cannot be
recorded due to some lack of resources. This fix adds a check to avoid
recording a command buffer which has a vkCmdExecuteCommands command
with as an argument a list of secondary command buffers where one of
the secondary command buffer recording has not been completed.

Note that the issue is not about the ordering of commandbuffer recordings:
the state rebuilder logic already makes sure to re-create all secondary
command buffers before primary ones. It is rather about secondary command
buffers which are, in the trace's initial state, marked as not terminated to
record: using any primary command buffer that refer to these secondary ones
would be invalid anyway, so it is not worth re-recording such unusable
command buffers.

Also, fix a variable name typo: use `cb` to refer to a command
buffer (`qp` probably came from a copy-paste of the loop on QueryPools
just above).

Bug: b/161071991

Test: manual, with a trace that had a vkCmdExecuteCommands with a
secondary command buffer whose recording was not completed.