Improve replay device selection dialog (#400)
Improve replay device selection dialog by displaying all known devices
in two tables: compatible devices which can be used for replay, and
incompatible devices, alongside the incompatibility reason to explain
for each why it is not fit to replay the trace.

This change modifies the semantics of GetDevicesForReplay(): it used
to return only compatible devices, now it returns a list of devices with
compatible devices at the front, and then incompatible devices. Two
companion lists indicates whether the device is compatible, and if not
the reason why it is incompatible (as a Stringtable.Msg).

On the GAPIS side, the incompatibility reason is set in
gapis/api/vulkan.GetReplayPriority(), and propagated to the answer to
a GetDevicesForReplay() RPC response.

On the GAPIC side, models.Devices.loadReplayDevices() handles the
augmented RPC reply and store replay device details in the new
ReplayDeviceInfo class. Compatible and incompatible devices are
each presented in a table in views.DeviceDialog. Selecting a compatible
replay device triggers a validation check, and if validation pass then the
user can open the trace with the selected device.

Bug: b/156719023
Test: manual