Add a mean to retrieve subCmd in MutateSubcommands (#514)
This change propagates a reference to the subcommand upon which
preSubCmdCallback() and postSubCmdCallback() are called.

Each of these two callbacks, used in MutateSubcommands() and
MutateWithSubcommands(), currently receives as argument:

- the complete api.SubCmdIdx of the subcommand

- the api.Cmd corresponding to the **parent** command

While there are ways to retrieve the actual subcommand from these two
arguments, it would be often a lot of work.

However, these callbacks are actually called in a context where we do
have a vulkan.CommandReferenceʳ to the subcommand, see
e.g. gapis/vulkan/api/externs.go:onPostSubcommand(). This reference is
being passed as an interface{} because it is API-specific, while
MutateSubcommands() is API-agnostic. This change makes sure that this
command reference is made accessible to the callbacks.

Currently, the only user of MutateWithSubcommands() is
FramebufferChangesResolvable.Resolve(), and it does not need this
command reference, but the upcoming framegraph code will require it.

Bug: b/171719615