Fix propagation of replay fatal error notifications (#560)
The replayer can send replay notifications to GAPIS, and the handling
of these notification can be customized for different kinds of
replays (e.g. report replays collect validation errors this way).

However, when a fatal error message notification is sent by the
replayer (e.g. because it has failed to create a Vulkan instance),
GAPIS should always treat this as an error, independently from the
custom notification handling.

In particular, this caused the frame profiling to hang when the replay
failed due to not being able to create a Vulkan instance.

Bug: b/172447447, b/173510056
Test: manual