Drop debug utils extension at replay time (#562)
At replay time, if Vulkan instance creation fail and we detect that
the validation layer or the debug report extension is requested, we
have a fallback where we try to create the instance again after
dropping the validation layer and the debug report extension. This
change adds the debug utils extension alongside the debug report
extension.

Note that both of these extensions are provided by the validation
layer. The typical failure is: the app requests the validation layer
at trace time, and one or both of the debug extensions. At replay
time, unless we do a report replay, the validation layer won't be
loaded, hence these extensions are not present.

More generally, we won't support replaying a workload that relies on
extensions provided by a layer, since this layer won't be present at
replay time. The validation layer is a special case since it is the
one layer we may provide at replay time (as it is present in our
APKs). See discussion at b/163754399

Bug: b/172447447
Test: manual