Framegraph: add coherent/mapped memory info (#585)
Buffers and images are backed by memory that may be coherent with the
host, and may be mapped at the time of their use. This change surfaces
this information in the framegraph's buffer and image types.

This turns out to be a rather big change because of some type
updates. The helpers lookup tables used to store references to Vulkan
BufferObject / ImageObject, from which we created framegraphBuffer /
framegraphImage later on. However, these objects do not store whether
the memory is coherent, for this we need to refer to the vulkan state
to access physical devices memory types. Therefore, this change
updates the lookup tables to store api.FramegraphBuffer /
api.FramegraphImage directly, where we can store the coherent memory
info. This info is retrievable when we update the lookup tables since
we have access to the vulkan state at this point.

A side-effect of storing api.FramegraphBuffer/api.FramegraphImages is
that for indexes of bufferAccesses/imageAccesses maps, we don't use
the typed VkBuffer/VkImage handles, but rather their uint64 value that
is stored in api.FramegraphBuffer/api.FramegraphImages. This also
triggers several trivial-but-noisy changes.

Also, document that readMappedCoherentMemory() results in a read-write
operation despite its name.

Bug: b/174471220
Test: manual