Simplify the apic_template rules to only generate code.
Instead of also trying to compile go code if any is generated. This
removes all the go toolchain deps when generating non-go code.

Also removed the deps attribute from the templates as it's no longer
used and dependencies need to be added to the rule that compiles the
generated code.