Replace MustRead by Read outside of generated code
MustRead should be called with a context that ignores
cancellation. That is the case in code generated for Mutate, outside
of this use-case we should use the regular Read and check for errors.

Note that code in externs.go is called from generated code, so it's OK
to use MustRead there.