Use RING_BUFFER fill policy for Perfetto buffers.
The DISCARD policy doesn't work as expected when writing into a file or
downloading the data via ReadBuffers. DISCARD does not allow the write
pointer of the buffer to wrap, even if the read pointer has been advanced
and so the maximum trace data is still going to be the size of the buffer.