Mock support of VkImages with Android external format (#781)
It is not possible to read directly from VkImages with an Android
external format (which are backed by an Android Hardware Buffer (AHB)
which has an external formal). To unblock the early analysis of some
apps using such kind of images, this change mocks support by actually
returning zeroed data for such images.

The proper way to retrieve such image data is to sample them, this is
tracked by b/148857112. However, there's likely other aspects to take
care of for properly re-creating such image data at replay time.

For now, this small change enables to unblock the early analysis of
some apps.

Bug: b/184956053