Add option to load validation layer at capture time on Android (#782)
This change adds a boolean "LoadValidationLayer" option to trace
options. When true, we load the Vulkan validation layer (VVL) below
our spy layer. This is useful for AGI developers to debug capture-time
issues.

This is only supported on Android, because we can rely on the VVL
which is present in our own APK. On desktop, we have no guarantee that
the VVL is available anywhere.

The const that defines the vulkan validation layer name is moved to
core/vulkan/loader package such that it can be easily accessed in both
vulkan transform and gapii client code.

Bug: n/a
Test: manual