Call adb unroot before installing apks (#786)
We're seeing a problem with apk installation on devices with adb root active. This unroots the adb server before trying to install APKs on device. I believe it should be safe because we explicitly call adb root before all places we need the permission, so temporarily revoking it shouldn't pose any danger. Hopefully!

You can test this by running "adb root" before trying to make a capture on a phone with a corp account.

Bug: 180757454