Disable UI support to edit shader source (#807)
Pushing the shader changes doesn't appear to work yet. Until we fix the
issue, hide the button and make the source code not editable.

This introduces a new flag to enable features that are under development. This
is intended for minor features that need to stay disabled while they are
implemented or issues are resolved.

Bug: 188434910
Test: manual