Fixes missing device id in Perfetto data. (#849)
Fixes missing device id in Perfetto data.

Sometimes Perfetto returns deviceID 0 for the QC devices which leads
to missing data.

There is no guaranteed way to fix this but for the Android we can assume
that, if there is only one device id is used for the entire trace,
it's very likely that missing deviceID is the same one rather than a
different device. If there are multiple devices than it does not do anything.