Fix screen auto-unlock for some devices (#868)
Replace the "menu" key event by the more robust "wm dismiss-keyguard"
command. In particular, this is the sequence used in various places in
the Android CTS, see:
https://cs.android.com/search?q=%22private%20void%20prepareDevice%22

Also, remove unreachable code at the end of UnlockScreen(): the
previous switch statement has a default case which always reach a
return, hence the return after this switch statement is effectively
unreachable (the code compiles fine without it).

Bug: b/193011219

Test: manual: take a trace on a device which has a locked screen,
using a device where auto-unlocking was failing before this change.