ABI by name lookup fixes.
Currently the ABIs' names need to be unique, because we lookup ABIs
by name. However, this lookup is only done in the Android parsing
code, so we can loosen this restriction to only require unique
names for the Android ABIs. This change also adds an init check to
ensure the Android ABI names are unique. This check would fail at
build time, as we use this during the codegen step.