Debug labels in the shader list view.
Turns the shader list from a tree into a table. It was a tree,
because when we had GLES support, shaders were grouped as children
below programs. Vulkan doesn't have programs, so the tree was
changed to be completely flat. Since this adds a second column,
change the tree to a table rather than a list.