Rework shader resource UX.
Show the shader data in the shader view as we used to, but allow shaders
to be pinned into their own tabs. This gets rid of the blank "Shader" tab,
but still allows viewing shaders in their own tabs. This also now allows
pinning shaders from the pipeline view.

Pinning is done via a new button in a new button bar in the shader UI.
Also adding functionality to export/save the shader code to a file, so
the pinning button isn't quite as lonely.

This adds a utility function in the theme to allow providing a size. The
resulting image will be a square of that size with the source image
centered. This basically allows adding padding to icons, so we can also
use these icons as buttons without adding a separate icon file with the
padding.