Handle log file creation errors in the client.
If the client fails to create the log file, it currently silently ignores
it. With this change, it will now attempt to create <logdir>/gapic.log,
followed by <logdir>/gapic-[0-9].log. If it fails in all cases, it gives
up and prints a message to stderr.

It now also attempts to create the specified log directory if it doesn't
already exist.