Pass the duration along in the trace config.
Makes sure the UI always populates either the number of frames to
capture for graphics traces, or the duration field for system
profiling traces.

Fixes the server side to push the handling of the duration down
into the tracer implementation, rather than trying to handle it
at the top. This makes deferred tracing with a duration work
again and removes the confusion with Perfetto tracing having
it's own implementation and requiring the duration to be zero.