Bring back the group tree for the profiling data.
This time mapping more closely to the command tree.

This makes some assumptions about command/sub command IDs:
1. we only get groups for command buffers, renderpasses and draw calls.
2. no overlaps.
3. the sub command ids are [cmdId, submission, cmdbuff, cmd].

All these assumptions currently hold and are also made in other parts of
the code in some way. The assumptions will need to be codified as part of
the command/sub-command refactor that is already planned.