Add VK_KHR_imageless_framebuffer (#1101)
Add feature and extension boilerplate. Capture and rebuild
VkFramebufferAttachmentsCreateInfo and VkRenderPassAttachmentBeginInfo
extension structs in vkCreateFramebuffer and vkCmdBeginRenderPass,
respectively. Track the now-dynamic framebuffer image attachments as
they are recorded (in CommandBufferObject) and submitted (in DrawInfo)
instead of in the FramebufferObject.

Tested with the imageless_framebuffer sample from
google/vulkan_test_applications