Add MEC support for invaliding command buffers.
This is the case where (for example) an image is destroyed, but
the image view is still referenced by a command buffer. This
is an invalidating operation, and means that we cannot re-recreate
the command buffer (nor do we have to MEC serialize it as it is
invalid at this point).